Add world-oriented properties to jsh.script
Also:
* Refine world-oriented jrunscript.file definitoins to remove methods
* Enhance type definitions internal to jsh.script
* Extract native Java type definitions to new source file
* Extract world-oriented jrunscript.file definitions to new source file
* Add two new world-oriented methods to jrunscript.file
* Move jrunscript/host Array definition from jsapi to Fifty
